backend/requirements.txt
amqp==5.0.5
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.5
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.6
django-cors-headers==3.7.0
django-redis==4.12.1
djangorestframework==3.12.2
importlib-metadata==3.4.0
isort==5.7.0
kombu==5.0.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==1.0.2
newrelic
# pkg-resources==0.0.0
prompt-toolkit==3.0.14
psycopg2-binary==2.8.6
pycodestyle==2.6.0
Pygments==2.7.4
pylint==2.6.0
pylint-django==2.4.2
pylint-plugin-utils==0.6
pytz==2021.1
redis==3.5.3
sentry-sdk==0.19.5
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
zipp==3.4.0
django-sendgrid-v5
